+++ python/source/CMakeLists.txt
@@ -14,12 +14,7 @@ SET(CPP_SOURCES
     python_wrappers.cc
     )
 
-
-ADD_LIBRARY(pycap MODULE ${CPP_SOURCES})
-SET_TARGET_PROPERTIES(pycap PROPERTIES 
-    PREFIX "_"
-    SUFFIX ".so"
-    )
+PYTHON_ADD_MODULE(pycap ${CPP_SOURCES})
 
 SET(PYTHON_INSTALL_PREFIX ${CMAKE_INSTALL_PREFIX}/lib/python${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}/site-packages)
 MESSAGE("PYTHON_INSTALL_PREFIX = ${PYTHON_INSTALL_PREFIX}")
