+++ python/source/__init__.py
@@ -4,7 +4,7 @@
 # without copyright and license information. Please refer to the file LICENSE
 # for the text and further information on this license. 
 
-from _pycap import *
+from .pycap import *
 from .data_helpers import *
 from .time_evolution import *
 from .end_criterion import *
@@ -14,12 +14,12 @@ from .voltammetry import *
 from .ragone_plot import *
 from .impedance_spectroscopy import *
 
-__all__ = ['_pycap', 'data_helpers', 'time_evolution', 'end_criterion',
+__all__ = ['pycap', 'data_helpers', 'time_evolution', 'end_criterion',
            'stage', 'charge_discharge', 'voltammetry', 'ragone_plot',
            'impedance_spectroscopy']
 
-__version__ = _pycap.__version__
-__git_branch__ = _pycap.__git_branch__
-__git_commit_hash__ = _pycap.__git_commit_hash__
-__git_remote_url__ = _pycap.__git_remote_url__
-__doc__ = _pycap.__doc__
+__version__ = pycap.__version__
+__git_branch__ = pycap.__git_branch__
+__git_commit_hash__ = pycap.__git_commit_hash__
+__git_remote_url__ = pycap.__git_remote_url__
+__doc__ = pycap.__doc__
