+++ python/source/python_wrappers.cc
@@ -28,7 +28,7 @@ build_energy_storage_device(boost::python::object & py_ptree,
     return cap::EnergyStorageDevice::build(comm, ptree);
 }
 
-BOOST_PYTHON_MODULE(_pycap)
+BOOST_PYTHON_MODULE(pycap)
 {
     if (import_mpi4py() < 0) return;
